+++ MastersOfMana/Assets/Scripts/Spells/SpellBehaviour/Summonings/FireballBehaviour.cs
@@ -108,7 +108,7 @@ public class FireballBehaviour : A_ServerMoveableSummoning
             {
                 directHitForce.Normalize();
                 directHitForce *= explosionForce;
-                player.movement.RpcAddForce(directHitForce, (int)ForceMode.VelocityChange);
+                player.movement.RpcAddForce(directHitForce, ForceMode.VelocityChange);
             }
         }
 
@@ -149,7 +149,7 @@ public class FireballBehaviour : A_ServerMoveableSummoning
 					force *= explosionForce;
 					Debug.DrawRay(c.attachedRigidbody.centerOfMass,force,Color.black,10);
 
-					ps.movement.RpcAddForce(force, (int)ForceMode.VelocityChange);
+					ps.movement.RpcAddForce(force, ForceMode.VelocityChange);
 				}
 				else
 				{
